Update PCB_testing_GPU_with_defect_extraction
This code takes a test image of a PCB and compares it pixel by pixel to the pre-defined reference image by running parallel threads on the GPU. Each thread compares the difference between the respective pixels of the reference and the test image. The program outputs the reference image, the test image with the defects boxed, the difference image and the components on the reference image which are detected as defects.