Update Image_blurring
This code takes a color image, separates it into red, green and blue channels, runs the kernel to blur an image using an averaging mask and then combines the channels back to form a color image. The code displays both the original image and the blurred version of the image.