Update PCB_testing_GPU
This code takes a test image of a PCB and compares it pixel by pixel to the pre-defined reference image by running parallel threads on the GPU. Each thread compares the difference between the respective pixels of the reference and the test image. The program outputs the reference image, the test image and the difference image.